重构工作流模型2.0

+++ 00_Source/01_Database/Database/Entity/Enums/Enums.cs
@@ -9,7 +9,7 @@ namespace Database.Entity.Enums
     public enum DBTYPE
     {
         INT,
-        INTEGER,
+        BIGINT,
 
         BIT,
         BOOLEAN,
